word2md/pyproject.toml
[project]
name = "word2md"
version = "0.1.0"
description = "Convert Word .docx files to Markdown"
requires-python = ">=3.11"
dependencies = [
    "python-docx>=1.1.0",
    "markdownify>=0.14.1",
]

[project.scripts]
word2md = "word2md.gui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
